requirements.txt
# Project dependencies
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
pyyaml>=6.0.1

# Testing dependencies
pytest>=8.0.0
pytest-mock>=3.14.0

# Optional but recommended
python-dotenv>=1.0.0  # for environment variable management
logging>=0.4.9.6      # for logging capabilities